requirements.txt
streamlit
pandas
snowflake-connector-python
google.generativeai
dotenv
faiss-cpu
sentence_transformers